requirements.txt
constriction
numpy
pandas
pillow
pytorch-msssim
PyYAML
tqdm
vector-quantize-pytorch
lpips
torchvision
jaxtyping
ninja
odak==0.2.6
pytorch_msssim